+++ src/main/java/org/example/util/CommonFunctionsUtil.java
@@ -3,7 +3,7 @@ package org.example.util;
 import java.util.Random;
 
 public class CommonFunctionsUtil {
-    public static boolean[] generateRandomPartition(int[] set, Random random) {
+    public static boolean[] generateRandomPartition(double[] set, Random random) {
         boolean[] partition = new boolean[set.length];
         for (int i = 0; i < set.length; i++) {
             partition[i] = random.nextBoolean();
@@ -11,9 +11,9 @@ public class CommonFunctionsUtil {
         return partition;
     }
 
-    public static int calculateDifference(int[] set, boolean[] partition) {
-        int sumA = 0;
-        int sumB = 0;
+    public static double calculateDifference(double[] set, boolean[] partition) {
+        double sumA = 0;
+        double sumB = 0;
         for (int i = 0; i < set.length; i++) {
             if (partition[i]) {
                 sumA += set[i];
